% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_consistent.R
\name{check_consistent}
\alias{check_consistent}
\title{Check for consistency of composite keys (handles '--' wildcards & aggregates output)}
\arguments{
\item{df}{Data frame to check}

\item{domain_name}{The domain of the dataset}

\item{target_vars}{Character vector of variables that should be unique}

\item{rule_id}{The rule ID being checked}

\item{group_vars}{Character vector of grouping variables. Default is NULL.}

\item{...}{Absorb extra parameters}
}
\description{
Check for consistency of composite keys (handles '--' wildcards & aggregates output)
}
\author{
Zhu Xiuling
}
