% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_date_logic.R
\name{check_date_logic}
\alias{check_date_logic}
\title{Check date logic comparison}
\usage{
check_date_logic(df, domain_name, target_vars, rule_id, operator, ...)
}
\arguments{
\item{df}{domain dataset to check}

\item{domain_name}{the domain name, used to replace "--" in target_vars with the actual domain code}

\item{target_vars}{a character vector of two variable names containing "--" placeholders, e.g., "--STDTC, --ENDTC"}

\item{rule_id}{the rule id to check}

\item{operator}{comparison operator: one of "<", "<=", ">", ">="}

\item{...}{additional arguments}
}
\value{
rule code, check status, error detail, row number of the error
}
\description{
SD0013、SD0025、SD1002、SD1331、SD1334、SD1335
}
\author{
Zhu Xiuling
}
