% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicate.R
\name{check_duplicate}
\alias{check_duplicate}
\title{Check for duplicate composite keys (handles '--' wildcards & aggregates output)}
\arguments{
\item{df}{Data frame to check}

\item{domain_name}{The domain of the dataset}

\item{target_vars}{Character vector of variables that should be unique}

\item{rule_id}{The rule ID being checked}

\item{group_vars}{Character vector of grouping variables. Default is NULL.}

\item{filter_values}{Optional value to filter the target variable before checking for duplicates. Default is NULL (no filtering).}

\item{...}{Absorb extra parameters}
}
\description{
Check for duplicate composite keys (handles '--' wildcards & aggregates output)
}
\author{
Zhu Xiuling
}
