% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sd0003.R
\name{check_sd0003}
\alias{check_sd0003}
\title{Check ISO 8601 date/time format}
\arguments{
\item{df}{domain dataset to check}
}
\value{
rule code, check status, error detail, row number of the error
}
\description{
SD0003: Value of Dates/Time variables (*DTC) must contain
valid date or valid date and time values and must conform to the
ISO 8601 international standard.
}
\author{
Zhu Xiuling
}
