% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sd0055.R
\name{check_sd0055}
\alias{check_sd0055}
\title{match variable types with SDTM}
\usage{
check_sd0055(df, domain_name)
}
\arguments{
\item{df}{domain dataset to check}

\item{domain_name}{the domain of the dataset}
}
\value{
rule code, check status, error detail, row number of the error
}
\description{
SD0055: Variable Data Types in the dataset should
match the variable data types described in SDTM.
}
\author{
Zhu Xiuling
}
