% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sd0056.R
\name{check_sd0056}
\alias{check_sd0056}
\title{check required variables}
\arguments{
\item{df}{domain dataset to check}

\item{domain_name}{the domain of the dataset}
}
\value{
rule code, check status, error detail, row number of the error
}
\description{
SD0056: Variables described in SDTM IG as Required
must be included in the dataset.
}
\examples{
check_sd0056(ae)
# minimal reproducible example
\dontrun{
  result <- check_sd0056(ae)
}
}
\author{
Zhu Xiuling
}
