% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sd1004.R
\name{check_sd1004}
\alias{check_sd1004}
\title{check variable length}
\arguments{
\item{df}{domain dataset to check}

\item{domain_name}{the domain of the dataset}
}
\value{
rule code, check status, error detail, row number of the error
}
\description{
SD1004: The value of Planned Arm Code (ARMCD)
should be no more than 20 characters in length.
}
\author{
Zhu Xiuling
}
