% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sd1009.R
\name{check_sd1009}
\alias{check_sd1009}
\title{check invalide ETCD value}
\usage{
check_sd1009(df, domain_name)
}
\arguments{
\item{df}{domain dataset to check}

\item{domain_name}{this is an empty parameter}
}
\value{
rule code, check status, error detail, row number of the error
}
\description{
SD1009: The value of Element Code (ETCD) should be
no more than 8 characters in length.
}
\examples{
check_sd1009(ta)
# minimal reproducible example
\dontrun{
  result <- check_sd1009(ta)
}
}
\author{
Zhu Xiuling
}
