% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sd1049.R
\name{check_sd1049}
\alias{check_sd1049}
\title{check variable length}
\arguments{
\item{df}{domain dataset to check}

\item{domain_name}{the domain of the dataset}
}
\value{
rule code, check status, error detail, row number of the error
}
\description{
SD1049: Qualifier Variable Label (QLABEL) value may have up to 40 characters.
}
\author{
Zhu Xiuling
}
