% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sd1259.R
\name{check_sd1259}
\alias{check_sd1259}
\title{check variable length}
\arguments{
\item{df}{domain dataset to check}

\item{domain_name}{the domain of the dataset}
}
\value{
rule code, check status, error detail, row number of the error
}
\description{
SD1259: The value of Set Code (SETCD) should be no more than 8 characters in length.
}
\author{
Zhu Xiuling
}
