% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sd1300.R
\name{check_sd1300}
\alias{check_sd1300}
\title{check variable length}
\arguments{
\item{df}{domain dataset to check}

\item{domain_name}{the domain of the dataset}
}
\value{
rule code, check status, error detail, row number of the error
}
\description{
SD1300: The value of DOMAIN should be no more than 2 characters in length, excluding Associated Persons domains.
}
\author{
Zhu Xiuling
}
