% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sd2001.R
\name{check_sd2001}
\alias{check_sd2001}
\title{check variable length}
\arguments{
\item{df}{domain dataset to check}

\item{domain_name}{the domain of the dataset}
}
\value{
rule code, check status, error detail, row number of the error
}
\description{
SD2001:The value of Actual Arm Code (ACTARMCD) should be no more than 20 characters in length.
}
\author{
Zhu Xiuling
}
