% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{domain_class_map}
\alias{domain_class_map}
\title{SDTM Domain and Class Mapping Table}
\format{
A data frame
}
\usage{
domain_class_map
}
\description{
Used to map specific domains (such as AE, DM, LB) to corresponding classes (such as Interventions, Events, Findings)
for obtaining rule IDs
}
\keyword{datasets}
