% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_rules.R
\name{fct_get_rules}
\alias{fct_get_rules}
\title{get the rule from P21}
\arguments{
\item{domain_name}{the name of the domain dataset, should be a character of two uppercase letters}
}
\value{
rule code
}
\description{
Extract the relevant rules from Pinnacle 21
based on the provided dataset
}
\author{
Zhu Xiuling
}
