% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_validate_sdtm.R
\name{fct_validate_sdtm}
\alias{fct_validate_sdtm}
\title{check the sdtm dataset}
\arguments{
\item{df}{domain dataset to check}
}
\value{
a validation report
}
\description{
validate the Submission-ready SDTM rds dataset,
including structure, values, and logic and output a report
}
\examples{
# minimal reproducible example
\dontrun{
  ta <- data.frame(
  Row = 1:6,
 STUDYID = c("EX7", "EX7", "EX7", "EX7", "EX7", "EX7"),
  DOMAIN = c("TA", "TA", "TA", "TA", "TA", "TA"),
  ARMCD = c(1, 1, 1, 1, 1, 1),
  ARM = c("CR", "CR", NA, "CR", "", "CR"),
  TAETORD = as.character(1:6),
  ETCD = c("SCRN", "ICR", "CRNS", "C", "C", "FU"),
  ELEMENT = c("Screen", "Initial Chemo + RT",
              "Chemo+RT (non-Surgery)", "Chemo",
              "Chemo", "Off Treatment Follow-up"),
  TABRANCH = c("Randomized to CR", "", "",
               "", "", ""),
  TATRANS = c("", "", "If progression, skip to Follow-up.",
              "", "", ""),
  EPOCH = c("SCREENING", "INDUCTION TREATMENT",
            "INDUCTION TREATMENT", "CONTINUATION TREATMENT",
            "CONTINUATION TREATMENT", "FOLLOW-UP")
)
  fct_validate_sdtm(ta)
}
}
\author{
Zhu Xiuling
}
