% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sdtm_v34}
\alias{sdtm_v34}
\title{SDTM v3.4 dictionary}
\format{
A data frame
}
\source{
CDISC SDTMIG v3.4 Standard Documentation.
}
\usage{
sdtm_v34
}
\description{
Includes basic mappings for the variables, data types, and controlled terms (CTs) defined in the SDTM v3.4 standard.
}
\keyword{datasets}
