% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_length.R
\name{check_length}
\alias{check_length}
\title{check the value length}
\usage{
check_length(df, domain_name, rule_id, variable_name, length_limit)
}
\arguments{
\item{df}{domain dataset to check}

\item{domain_name}{this is an empty parameter}

\item{rule_id}{the rule id to check}

\item{variable_name}{the variable name to check}

\item{length_limit}{the length limit for the variable}
}
\value{
rule code, check status, error detail, row number of the error
}
\description{
SD1004、SD1009、SD1022、SD1049、SD1231、SD1259、SD1300、SD1475、SD2001
}
\author{
Zhu Xiuling
}
