% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sd0002.R
\name{check_sd0002}
\alias{check_sd0002}
\title{check required variables}
\arguments{
\item{df}{domain dataset to check}

\item{domain_name}{the domain of the dataset}
}
\value{
rule code, check status, error detail, row number of the error
}
\description{
SD0002: Required variables (where Core attribute is 'Req')
cannot be null for any records.
}
\author{
Zhu Xiuling
}
