% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sd1009.R
\name{check_sd1009}
\alias{check_sd1009}
\title{check variable length}
\usage{
check_sd1009(df, domain_name)
}
\arguments{
\item{df}{domain dataset to check}

\item{domain_name}{the domain of the dataset}
}
\value{
rule code, check status, error detail, row number of the error
}
\description{
SD1009:The value of Element Code (ETCD) should be no more than 8 characters in length.
}
\author{
Zhu Xiuling
}
