% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sd1022.R
\name{check_sd1022}
\alias{check_sd1022}
\title{check variable length}
\usage{
check_sd1022(df, ...)
}
\arguments{
\item{df}{domain dataset to check}

\item{...}{Additional arguments (currently unused).}
}
\value{
rule code, check status, error detail, row number of the error
}
\description{
SD1022: The value of Qualifier Variable Name (QNAM)
should be limited to 8 characters, cannot start with a number,
and cannot contain characters other than letters in upper case, numbers, or underscores.
}
\author{
Zhu Xiuling
}
