% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sd1064.R
\name{check_sd1064}
\alias{check_sd1064}
\title{check duplicate ETCD value}
\usage{
check_sd1064(df, domain_name)
}
\arguments{
\item{df}{domain dataset to check}

\item{domain_name}{this is an empty parameter}
}
\value{
rule code, check status, error detail, row number of the error
}
\description{
SD1064: The value of Element Code (ETCD) variable
must be unique within Trial Elements (TE) domain.
}
\examples{
check_sd1064(ta)
# minimal reproducible example
\dontrun{
  result <- check_sd1064(ta)
}
}
\author{
Zhu Xiuling
}
