% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_rules.R
\name{fct_get_rules}
\alias{fct_get_rules}
\title{get the rule from P21}
\arguments{
\item{df}{domain dataset to check}
}
\value{
rule code
}
\description{
Extract the relevant rules from Pinnacle 21
based on the provided dataset
}
\examples{
get_rules(ta)
# minimal reproducible example
\dontrun{
  result <- get_rules(ta)
}
}
\author{
Zhu Xiuling
}
