% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_validate_sdtm.R
\name{fct_validate_sdtm}
\alias{fct_validate_sdtm}
\title{check the sdtm dataset}
\usage{
fct_validate_sdtm(df)
}
\arguments{
\item{df}{domain dataset to check}
}
\value{
a validation report
}
\description{
validate the Submission-ready SDTM rds dataset,
including structure, values, and logic and output a report
}
\examples{
fct_validate_sdtm(ta)
# minimal reproducible example
\dontrun{
  fct_validate_sdtm(ta)
}
}
\author{
Zhu Xiuling
}
