% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{generate a html report}
\usage{
generate_report(report_df, domain_name, output_dir = getwd())
}
\arguments{
\item{df}{a data.frame of all errors detail}

\item{the}{domain of the dataset}
}
\value{
a html report
}
\description{
SD0002: Required variables (where Core attribute is 'Req')
cannot be null for any records.
}
\author{
Zhu Xiuling
}
