% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_error.R
\name{report_error}
\alias{report_error}
\title{a error per line}
\usage{
report_error(row_number, variable_name, original_value, rule_id, error_message)
}
\arguments{
\item{row_number}{specific error row number}

\item{variable_name}{specific error variable name}

\item{original_value}{the original value of the error}

\item{rule_id}{error rule ID}

\item{error_message}{the description of the error}
}
\value{
a date.frame
}
\description{
Summary of results
}
\author{
Zhu Xiuling
}
